# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.10.0
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    tr-engine authenticates **client software, not people**. Every client
    (a dashboard deployment, a script, an upload plugin, a website's
    backend, a phone app) holds its own **API key** and sends it as:

    ```
    Authorization: Bearer tre_<64 hex characters>
    ```

    There are no user accounts, logins, sessions or cookies. How a client
    authenticates its own users, if it has any, is up to the client. The
    operator guide and the client-developer guide are in `docs/auth.md`.

    **Scopes.** A key has one of `listen`, `edit` or `admin`, optionally
    plus `upload`, or only `upload`:

    | Scope | Grants | Implies |
    |---|---|---|
    | `listen` | Read radio data, stream events and audio, mint tickets | |
    | `edit` | Change radio metadata (tags, transcription corrections and reviews, unit-tag suggestions) | `listen` |
    | `admin` | Everything else: keys, anonymous access, identity edits, merges, maintenance, storage, SQL query, pages, imports, console logs, audit log | `edit`, `listen` |
...

...
    what applies to it.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /audio/jitter:
    get:
      operationId: getAudioJitter
      x-scope: listen
      x-restricted: deny
      summary: Get audio jitter stats
      description: |
        Returns per-stream inter-packet arrival jitter statistics for all
        active audio streams. Jitter is measured using Welford's online
        algorithm on inter-chunk arrival deltas at the server's UDP receive
        layer.

        > **Note:** This feature is only available when `STREAM_LISTEN` is
        > configured. Returns 404 when not enabled.
      tags: [Audio]
      responses:
        "200":
          description: Per-stream jitter statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  streams:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/StreamJitterSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Live audio streaming is not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Pages (Web UI metadata)
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: tre_<64 hex characters>
      description: |
        An API key in the `Authorization` header:
        `Authorization: Bearer tre_...`. Keys are created by an admin
        (`POST /keys` or `tr-engine keys create`). Keys are never accepted
        from the URL.
    ticketAuth:
      type: apiKey
      in: query
      name: ticket
      description: |
        A short-lived, listen-only ticket from `POST /tickets`, passed as
        `?ticket=trt_...`. Accepted only by `GET`/`HEAD` on
        `/events/stream`, `/audio/live` and `/calls/{id}/audio`, for
        browser APIs that cannot send headers (`EventSource`,
        `<audio src>`, `WebSocket`). Takes precedence over an
        `Authorization` header on those operations.

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
...
  responses:
...
    Unauthorized:
      description: |
        No usable credential. `code` is `key_required` (no credential, and
        the anonymous access policy doesn't allow this operation),
        `invalid_key` (unknown, revoked or expired key) or `invalid_ticket`
        (malformed, expired or orphaned ticket).
      headers:
        WWW-Authenticate:
          description: Always `Bearer realm="tr-engine"`
          schema:
            type: string
            example: Bearer realm="tr-engine"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            key_required:
              value:
                code: key_required
                error: an API key is required
            invalid_key:
              value:
                code: invalid_key
                error: API key has been revoked
    Forbidden:
      description: |
        The credential is valid but not allowed here. `code` is
        `insufficient_scope` (the message names the scope the operation
        needs) or `restricted_credential` (the credential is restricted to
        some systems or talkgroups, and this operation can't enforce that).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficient_scope:
              value:
                code: insufficient_scope
                error: this operation needs the edit scope
            restricted_credential:
              value:
                code: restricted_credential
                error: this operation is not available to restricted credentials
...
  schemas:
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
            - key_required
            - invalid_key
            - invalid_ticket
            - insufficient_scope
            - restricted_credential
          description: |
            Machine-readable error code for programmatic handling. Auth codes:
            - `key_required` (401): no credential, and the anonymous access
              policy doesn't allow this operation
            - `invalid_key` (401): unknown, revoked or expired key
            - `invalid_ticket` (401): malformed, expired or orphaned ticket,
              or one that references a merged-away system
            - `insufficient_scope` (403): the key lacks the needed scope
            - `restricted_credential` (403): the credential is restricted and
              this operation can't enforce that
            - `service_unavailable` (503): also when the credential lookup
              failed
            - `forbidden` (403): from the auth layer, only for a route with no
              auth policy (a server bug)
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    StreamJitterSnapshot:
      type: object
      properties:
        system_id:
          type: integer
          description: System ID
        tgid:
          type: integer
          description: Talkgroup ID
        count:
          type: integer
          description: Number of inter-packet deltas recorded
        min_ms:
          type: number
          format: double
          description: Minimum inter-packet delta in milliseconds
        max_ms:
          type: number
          format: double
          description: Maximum inter-packet delta in milliseconds
        mean_ms:
          type: number
          format: double
          description: Mean inter-packet delta in milliseconds
        stddev_ms:
          type: number
          format: double
          description: Standard deviation of inter-packet deltas in milliseconds
        last_delta_ms:
          type: number
          format: double
          description: Most recent inter-packet delta in milliseconds
        started_at:
          type: string
          format: date-time
          description: When this stream became active
...